# backend/internal/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Improver API",
    "version": "0.1.24"
  },
  "paths": {
...
    "/api/v1/admin/sources/{source_id}/mts-link/organizations": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Organizations",
        "operationId": "organizations_api_v1_admin_sources__source_id__mts_link_organizations_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SsoEmail"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}/mts-link/login-email": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Login Email",
        "operationId": "login_email_api_v1_admin_sources__source_id__mts_link_login_email_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}/mts-link/start": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Start",
        "operationId": "start_api_v1_admin_sources__source_id__mts_link_start_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SsoStart"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sources/{source_id}/mts-link/finish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Finish",
        "operationId": "finish_api_v1_admin_sources__source_id__mts_link_finish_post",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SsoFinish"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SourceRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "settings": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagReference"
            },
            "type": "array",
            "title": "Tags"
          },
          "credential_configured": {
            "type": "boolean",
            "title": "Credential Configured"
          },
          "refresh_token_configured": {
            "type": "boolean",
            "title": "Refresh Token Configured",
            "default": false
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "source_type",
          "enabled",
          "settings",
          "tags",
          "credential_configured",
          "last_sync_at",
          "last_error",
          "created_at",
          "updated_at"
        ],
        "title": "SourceRead"
      },
...
      "SsoEmail": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 3,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "SsoEmail"
      },
      "SsoFinish": {
        "properties": {
          "ticket": {
            "type": "string",
            "maxLength": 65536,
            "format": "password",
            "title": "Ticket",
            "writeOnly": true
          },
          "auth_code": {
            "type": "string",
            "maxLength": 32768,
            "format": "password",
            "title": "Auth Code",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "ticket",
          "auth_code"
        ],
        "title": "SsoFinish"
      },
      "SsoStart": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 320,
            "minLength": 3,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Email"
          },
          "organization_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Organization Id"
          },
          "method_index": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Method Index"
          },
          "enable_source": {
            "type": "boolean",
            "title": "Enable Source",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "email",
          "organization_id",
          "method_index"
        ],
        "title": "SsoStart"
      },
...
      "TagReference": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TagReference"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}